archive/hhayan/src_test/requirements.txt
# requirements.txt
python-dotenv==1.1.1
langgraph==0.6.7
langchain==0.3.27
langchain-core==0.3.76
langchain-openai==0.3.33
langchain-text-splitters==0.3.11
langchain-community==0.3.27
langchain-pinecone==0.3.11
pinecone==5.3.1
openai==1.107.3
